move out the function to get the list of wizard links
